+++ tests/test_config.py
@@ -92,7 +92,7 @@ class TestExampleProject(TestCase):
             self.assertEqual(settings.SYSTEM_ASSETS, [])
             self.assertFalse(settings.DRAFT_MODE)
             self.assertTrue(settings.CACHE_REBUILDS)
-            self.assertEqual(settings.TAXONOMIES, ["tags", "series"])
+            self.assertEqual(settings.TAXONOMIES, [])
 
     def _test_project(self, root, title='Example web site', cfg='', **kw):
         args = test_utils.Args(project=root + cfg, **kw)
@@ -105,6 +105,8 @@ class TestExampleProject(TestCase):
         with open(output, "rt") as fd:
             content = fd.read()
 
+        # FIXME: this does not make sense anymore, since title is not defined
+        # by settings anymoer
         self.assertIn(title, content)
 
     def test_default_name(self):