+++ tests/utils.py
@@ -28,7 +28,8 @@ class Site(staticsite.Site):
     def load_without_content(self):
         self.features.load_default_features()
         self.load_theme()
-        self.theme.load_assets()
+        with tempfile.TemporaryDirectory() as dir:
+            self.load_content(content_root=dir)
 
 
 def datafile_abspath(relpath):